+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-# initial
+## v1.0.0
+- feature: #2 Implement balance endpoint
+   + Add Wallet Entity
+   + Add Wallet Repository
+   + Add Wallet Service
+   + Date Persistance Configuration
+   + Add Wallet Balance Test
+   + Add Rest API
+
+
+## initial
 - #1 Define Project Scope and Initial Requirements.
    - #1 Update ProblemStatment.md and README.md
    - #1 Define main features and MVP